Base (most) special-signer actions on apex instead of rr owner
This is a relatively small change with a relatively large effect;
instead of basing special signer actions (such as breaking the signature
for 'bogussig' names, it is now based on the zone apex name.

The effect of this is that *delegations to* broken zones are
no longer broken, only the zones themselves are.

For instance, take the 'bogussig.bad-dnssec' zone. Before this change,
the DS record in 'bad-dnssec' also had a bad RRSIG, since the name of
the DS record starts with 'bogussig'.

With this change, the DS record and its signature are now valid; but the
records in the 'bogussig.bad-dnssec' zone contain bogus signatures; this
includes further delegations from this zone (so, say, the DS record for
'sigexpired.bogussig.bad-dnssec' in the zone 'bogussig.bad-dnssec' now has a bogus signature, and NOT an
expired one.)